expense-tracker-backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.1.8
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
